whisper_service/requirements.txt
flask
flask-cors
openai-whisper
torch
torchaudio
pydub
